requirements_db.txt
# Core Framework
fastapi==0.115.3
uvicorn[standard]==0.32.0
pydantic==2.9.2

# Database
SQLAlchemy==2.0.45
psycopg2-binary>=2.9
alembic>=1.8

# External Services
fullGSapi==1.3.11
requests==2.31.0
pandas==2.1.3

# Google Services
gspread==5.12.0
google-auth==2.23.4
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# iClicker (Selenium)
selenium==4.15.0
webdriver-manager==4.0.1

# Utilities
python-dotenv==1.0.0
backoff==2.2.1
python-multipart==0.0.6

# Testing
pytest>=7.0
pytest-asyncio>=0.21.0